--- src/components/nrc-content/indicators/indicators.js
@@ -1,6 +1,0 @@
-import { connect } from 'react-redux';
-
-import Component from './indicators-component';
-import mapStateToProps from './indicators-selectors';
-
-export default connect(mapStateToProps, null)(Component);
