+++ src/components/nrc-content/nrc-challenges/nrc-challenges-constants.js
