+++ src/components/nrc-content/nrc-challenges/nrc-challenges-selectors.js
