+++ src/components/nrc-content/nrc-footer/nrc-footer-styles.module.scss
