+++ src/components/nrc-content/nrc-indicators/indicator-card/indicator-card-component.jsx
