+++ src/components/nrc-content/nrc-indicators/nrc-indicators-selectors.js
