+++ src/components/nrc-content/nrc-indicators/nrc-indicators-styles.module.scss
