+++ src/components/nrc-content/nrc-indicators/nrc-indicators-utils.js
