+++ src/components/nrc-content/nrc-trend/area-chart-tooltip/area-chart-tooltip-component.jsx
