+++ src/components/nrc-content/nrc-trend/nrc-trend-styles.module.scss
