+++ src/components/nrc-content/nrc-vertebrates/nrc-vertebrates.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { connect } from 'react-redux';
 import uiActions from 'redux_modules/ui';
 
-import Component from './vertebrates-component';
-import mapStateToProps from './vertebrates-selectors';
+import Component from './nrc-vertebrates-component';
+import mapStateToProps from './nrc-vertebrates-selectors';
 
 const actions = {
   ...uiActions,
