+++ src/components/pdf-reports/national-report-pdf/national-report-pdf-component.jsx
@@ -7,8 +7,8 @@ import { getLocaleNumber } from 'utils/data-formatting-utils';
 
 import AreaChart from 'components/charts/area-chart';
 import HalfEarthLogo from 'components/half-earth-logo';
-import IndicatorCard from 'components/nrc-content/indicators/indicator-card';
-import { getBarStyles } from 'components/nrc-content/indicators/indicators-utils';
+import IndicatorCard from 'components/nrc-content/nrc-indicators/indicator-card';
+import { getBarStyles } from 'components/nrc-content/nrc-indicators/nrc-indicators-utils';
 
 import { LAND_MARINE } from 'constants/country-mode-constants';
 import { getCountryNames } from 'constants/translation-constants';
